+++ android/app/src/main/res/values/arrays.xml
@@ -3,6 +3,5 @@
     <string-array name="snap_connect_scopes">
         <item>https://auth.snapchat.com/oauth2/api/user.bitmoji.avatar</item>
         <item>https://auth.snapchat.com/oauth2/api/user.display_name</item>
-        <item>https://auth.snapchat.com/oauth2/api/user.external_id</item>
     </string-array>
 </resources>
